Use `workspaces` manifest entry to find workspaces
The `workspaces` property in `package.json` is now used to find
workspaces. Previously it was assumed that packages were in the
directory 'packages'.

The package `glob` has been added to parse the `workspaces` entries.
This is the package that Yarn itself uses [1], so it should work well
enough here.

As part of this change, the `dirPath` property in the package metadata
we pass around has been changed to represent the package path relative
to the root directory. Previously it was the either absolute or
relative to the current working directory. This was easier, and it
let us delete the utility that was used for shortening the path to a
readable length (which was itself assuming the use of a packages
directory).

[1]: https://github.com/yarnpkg/yarn/blob/3119382885ea373d3c13d6a846de743eca8c914b/src/config.js#L821