Add require-additional-reviewer workflow (#72)
This PR adds a GitHub Actions workflow named `require-additional-reviewer`, that uses an external action called `MetaMask/action-require-additional-reviewer`. This workflow requires that release PRs created by `action-create-release-pr` are reviewed by at least one organization member other than the release author. `action.yml` in this repository was modified to upload an artifact identifying the GitHub user that initiated it, which is then used in `require-additional-reviewer.yml`. This new workflow will fail if used until the new external action is published, but it can be ignored until such time.