Create build-only tsconfig
These changes are copied from the [module template][1]. The main goal
here is to expand TypeScript's purview so that it includes test files,
improving the overall development experience. We also remove `typeRoots`
from the config (because the value provided is already the default) and
add `exclude`s in the default `tsconfig.json` (for consistency with the
module template).

[1]: https://github.com/MetaMask/metamask-module-template/pull/102