Transpile module syntax into CJS in dev
`src/index.ts` is designed to be run as an executable, and we want to be
able to run this in tests via `ts-node`. However, `ts-node` does not run
a TypeScript file when `tsconfig.json` specifies anything other than
"CommonJS" for the `module` setting ([unless you configure `ts-node`
and/or Node specially][1]). This commit instructs TypeScript to rewrite
imports using the CommonJS syntax only in the default TypeScript config
(ES6 is still used when `yarn build` is run).

[1]: https://typestrong.org/ts-node/docs/imports/