Bump Jest to 28.x and bump TypeScript to 4.3.0
Because Jest no longer uses Jasmine, but its own custom test runner
called Circus, this fixes a bug in Jest where if an error occurs in a
`beforeAll` hook it does not prevent the tests that make use of the hook
from running.

This commit also bumps `ts-jest`, which compels us to bump TypeScript as
well.

[1]: https://github.com/facebook/jest/issues/2713