Add high-level tests
This commit adds tests which exercise the behavior of the script that
powers the majority of this action. It does this by using real Git repos
and testing the contents of real files instead of mocking out commands
and access to the filesystem.

The tests in this commit should cover all of the happy paths provided
directly by this action, as well as some behavior offered by
`@metamask/auto-changelog` is also covered, because it seemed to make
sense when thinking about how this action "should" work from a high
level.