Add JSDoc linting rules, update lint deps
`eslint-config-jsdoc` was not listed as one of the dev dependencies, so
`yarn lint` wasn't checking anything around JSDoc. This commit adds this
dependency and fixes JSDoc blocks that fail the linter. It also updates
all of the existing `eslint-config-*` dependencies to their latest
versions and fixes linting issues that arise because of this.