Retry getting the release PR number on a timeout (#96)
Recently, this action [failed](https://github.com/MetaMask/post-message-stream/runs/7096918602?check_suite_focus=true) because we failed to get the release PR number via `gh pr view`, despite the fact that the PR was created successfully. I hypothesize that this happened because we didn't give GitHub enough time to create the PR and return its number to us. This PR ensures that we retry `gh pr view` on a 10-second interval for a total of 60 seconds, to give GitHub enough time to return the result we're looking for.