Check for `workspace:^` before replacing the version in package.json (#108)
* Check for presence of `workspace:^` in package.json before replacing version

This change modifies getUpdatedDependencyField function to check the dependencyObject values for `workspace:^`
as a condition before replacing the version with newVersionRange. The 'workspace:^' should not be overridden with
newVersionRange.

* Add a step to use Node.js version specified in .nvmrc

Without specifying a version in the workflow, the job could fail or use the wrong version of Node.js. This change ensures that the job uses the Node.js version specified in .nvmrc.

* Update build files

* Check if dependency starts with 'workspace:'

This change checks whether a dependency field starts with 'workspace:', instead of being equal to 'workspace:^'.